Bug [1917971] fractional CR are not represented properly in PCGen
git-svn-id: svn+ssh://svn.code.sf.net/p/pcgen/code/Trunk@5645 ef53c05b-3124-4d59-ae7d-fe3273f42be7

+++ code/src/java/plugin/lsttokens/template/CrToken.java
@@ -18,7 +18,7 @@ public class CrToken implements PCTemplateLstToken
 	{
 		try
 		{
-			template.setCR(Integer.parseInt(value));
+			template.setCR(Float.parseFloat(value));
 		}
 		catch (NumberFormatException nfe)
 		{
